requirements.txt
pykakasi==2.2.1
PyYAML==6.0.1
tqdm==4.65.0
